Fix 4 more re-truncated DAT_000bbf70 pointers + a stack-pointer-arg truncation in the babl symbol-table cluster
The previous Talk-crash fix (d1525ac) widened DAT_000bbf70 itself to
intptr_t and widened FUN_0001b0a4's own local copy of it to match, but
missed 4 sibling readers of the same cluster that still declared their
local copy as plain `int`: FUN_0001ac48, FUN_0001acf8, FUN_0001aebc,
FUN_0001afe4. Confirmed live via lldb against bug-critter-talk.txt,
which exercises a genuine successful CNV.ARK conversation load (Bragit
does have real dialogue) rather than the "no record" early-out the
first fix chased -- so DAT_000bbf70 is non-NULL here and these
functions' own truncation of it is what crashes, one step further into
FUN_0002a8e0's npc_whoami/npc_hunger/etc. builtin-variable lookups.

Also found and fixed a second, unrelated truncation in the same two
functions: FUN_0001aebc/FUN_0001afe4's own `param_2` was declared plain
`int`, but every real caller passes a stack pointer (e.g. FUN_0002a8e0's
`local_20`, FUN_00019e58's own siblings' `&local_2c`/`&local_10`) --
these are old K&R-style definitions with no prototype to catch the
mismatch, so the pointer silently truncated to 32 bits and the
function's own dereference of it crashed next (confirmed live: fault
address 0x6fdfe010 is the real ~0x1_6fdfe010 stack address with its
top byte dropped).

bug-critter-talk.txt (talking to an NPC with a real conversation) and
demo_critter_talk_test.txt (Bragit, no record) both now run clean; full
demo_*.txt regression suite still clean.

+++ uw.c
@@ -12078,8 +12078,8 @@ void FUN_0001ac48()
   int iVar1;
   int iVar2;
   int iVar3;
-  int iVar4;
-  
+  intptr_t iVar4; // was `int` -- re-truncated DAT_000bbf70 (now intptr_t) right back down, same as FUN_0001b0a4's own fix
+
   iVar1 = FUN_0007863c((int)*(short *)(DAT_000bbf0c + DAT_000bbf78 * 2));
   iVar2 = FUN_00019aa0();
   DAT_000bbf78 = DAT_000bbf78 + -1;
@@ -12112,8 +12112,8 @@ void FUN_0001acf8()
   int iVar1;
   int iVar2;
   int iVar3;
-  int iVar4;
-  
+  intptr_t iVar4; // was `int` -- re-truncated DAT_000bbf70 (now intptr_t) right back down, same as FUN_0001b0a4's own fix
+
   iVar1 = FUN_0007863c((int)*(short *)(DAT_000bbf0c + DAT_000bbf78 * 2));
   iVar2 = FUN_00019aa0();
   DAT_000bbf78 = DAT_000bbf78 + -1;
@@ -12232,18 +12232,18 @@ undefined4 param_2;
 
 void FUN_0001aebc(param_1,param_2,param_3)
 char *param_1;
-int param_2;
+intptr_t param_2; // was `int` -- every real caller passes a stack pointer (e.g. FUN_0002a8e0's `local_20`), truncated on 64-bit; same bug class as DAT_000bbf70 (crashes at param_2's own dereference, uw.c ~12283)
 short param_3;
 
 {
   undefined1 uVar1;
-  int iVar2;
+  intptr_t iVar2; // was `int` -- re-truncated DAT_000bbf70 (now intptr_t) right back down, same as FUN_0001b0a4's own fix; this is the crash in bug-critter-talk.txt's own successful-conversation-load path (via FUN_0002a8e0's npc_whoami lookup)
   uint uVar3;
   int iVar4;
   uint uVar5;
   short sVar6;
   undefined1 local_34 [28];
-  
+
   sVar6 = 0;
   iVar2 = Ordinal_1068();
   if (iVar2 != 0) {
@@ -12290,13 +12290,13 @@ short param_3;
 
 void FUN_0001afe4(param_1,param_2,param_3)
 char *param_1;
-int param_2;
+intptr_t param_2; // was `int` -- same pointer-truncation bug as FUN_0001aebc's own param_2 (every real caller passes a stack pointer, e.g. `&local_10`)
 short param_3;
 
 {
   int iVar1;
-  int iVar2;
-  
+  intptr_t iVar2; // was `int` -- re-truncated DAT_000bbf70 (now intptr_t) right back down, same as FUN_0001b0a4's own fix
+
   iVar2 = DAT_000bbf70;
   while( true ) {
     /* Same DAT_000bbf70-uninitialized guard as babl_register_builtin's own
